:sparkles: feat : 컴포 넌트 분리 #3
