+++ client/data/data.json
@@ -20,8 +20,9 @@
       "name": "홍길동",
       "comment": "good",
       "avatar": "../src/assets/avatar/1.jpg",
-      "value": "3",
+      "value": 3,
       "Good": 5,
+      "createdAt": "2022. 11. 18. AM 10:40",
       "photo": [
         {
           "img": "../src/assets/food/ramen.jpg",
@@ -42,7 +43,9 @@
       "name": "고길동",
       "comment": "bad",
       "avatar": "../src/assets/avatar/1.jpg",
-      "value": "4",
+      "value": 4,
+      "Good": 3,
+      "createdAt": "2022. 11. 19. AM 12:40",
       "photo": [
         {
           "img": "../src/assets/food/ramen.jpg",
@@ -62,7 +65,9 @@
       "id": 3,
       "name": "김길동",
       "avatar": "../src/assets/avatar/1.jpg",
-      "value": "2",
+      "value": 2,
+      "Good": 1,
+      "createdAt": "2022. 11. 20. AM 12:34",
       "comment": "SoSo"
     }
   ]
