+++ client/src/components/Items/HeaertVoteButton.jsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import FavoriteIcon from '@mui/icons-material/Favorite';
 import ToggleButton from '@mui/material/ToggleButton';
 
-export default function StandaloneToggleButton() {
+export default function HeaertVoteButton() {
   const [selected, setSelected] = React.useState(false);
 
   return (
